feat(io_uring): fix ws bugs (#2279)
Bugs fixed:

- TLS handshake failed: This is due to system scenario test expecting
plan WS server and in the config TLS for WS was enabled.
- Connection closed log noise: This was happening because the websocket
client in benchmark sends `Close` frame to server and drops the
connection without waiting for ack from server. Due to this server
throws `Io error BrokenPipe`. This is expected in real world usage and
we handle this on the server side and downgrade the specific errors to
warning.